feat(navbar): toggle sidebar on active main nav item click
Adds the ability to toggle the associated sidebar by clicking on an
already active main navigation item.

If the item's sidebar is open, it will be closed. If it is closed,
it will be opened as a floating sidebar. This provides a quick way
to hide and show sidebars without deselecting the main view.